Explore:Transferred calibration & iris detection to Misc Activity
 Explore:
1.transferred the claibration & iris detectio so to make the main screen less bloated
