+++ Explore/app/src/main/res/values/strings.xml
@@ -15,11 +15,13 @@
     <string name="eye_gaze_blob">Eye gaze blob</string>
     <string name="eye_gaze_shape">Eye gaze shape</string>
     <string name="eye_gaze_max_area">Eye gaze max area</string>
-    <string name="eye_gaze_sparse_optical_flow">Eye gaze Sparse Optical Flow</string>
+    <string name="eye_gaze_optical_flow">Eye gaze Optical Flow</string>
     <string name="settings_name">Settings</string>
     <string name="re_calibrate">Re-Calibrate</string>
     <string name="left">Left</string>
     <string name="right">Right</string>
+    <string name="eye_gaze_optical_flow_ui">Eye Gaze Optical Flow UI</string>
+    <string name="calibration">Calibration</string>
 
 
 </resources>
